backend/requirements.txt
Flask
Flask-Cors==3.0.10
beautifulsoup4==4.9.3
requests==2.32.3
googlesearch-python==1.2.4
networkx==3.3
langchain==0.2.11
langchain-community==0.2.10
pandas==2.2.2
numpy==1.26.4
pathlib==1.0.1
yachalk==0.1.6
tqdm==4.66.5
unstructured==0.15.0
python-magic-bin==0.4.14